mon: AuthMonitor: Don't always purpose full versions
We cannot purpose a full version if the KeyServer has no keys. This should
only happen when we start with a brand new store, and at the time we are
creating the first pending proposal.

However, committing a full version full of nothingness creates a big
problem, since we won't care for the incrementals if we have a full version
on store -- which we would then have, however corrupted it was.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>