mon: Monitor: keep track of paxos version when starting a store sync
We were ignoring at which version a store sync started given that after
the sync we end up with the same state as the monitor we are sync'ing
from, which should let us end up with a consistent state.

However, we were incurring in two mistakes:

- initializing the paxos namespace iterator *after* we had exhausted the
whole-store iterator, thus potentially capturing a different state of the
store at that time; and,

- missing that the leader stashes the current proposal prior to committing
it (if any is on-going).

(to be continued after rebasing against the patch that should enfore paxos
to reapply all its versions after sync)

Fixes: #4026
Fixes: #4037
Fixes: #4040

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>