mon: Paxos: 'lock' paxos when we are running some internal proposals
Force the paxos services to wait for us to complete whatever we are doing
before they can proceed.  This is required because on certain occasions
we might need to run internal proposals, not affected to any of the paxos
services (for instance, when learning an old value), and we need them to
stay put, or they might incur in erroneous state and crash the monitor.

Fixes: #4175

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>